+++ interfaces/EnumInterface.php
@@ -1,0 +1,7 @@
+<?php
+interface EnumInterface
+{
+    public function filter(mixed $data, ...$params): mixed;
+    public function format(mixed $data, ...$params): mixed;
+    public function validate(mixed $data, ...$params): void;
+}