+++ interfaces/FidProviderInterface.php
@@ -3,6 +3,6 @@ namespace Flex\Banana\Interfaces;
 
 interface FidProviderInterface
 {
-    public function getTable(): string;
-    public function getFidColumnName(): string;
+    public function getTable(?string $table): string;
+    public function getFidColumnName(?string $columnName): string;
 }